+++ ios/Classes/arabic_text_justification.h
@@ -1,0 +1,6 @@
+#ifndef ARABIC_TEXT_JUSTIFICATION_IOS_H
+#define ARABIC_TEXT_JUSTIFICATION_IOS_H
+
+#include "../../src/harfbuzz_wrapper.h"
+
+#endif