+++ lib/arabic_text_justification.dart
@@ -1,3 +1,4 @@
 export 'src/models.dart';
 export 'src/painter.dart';
 export 'src/arabic_text_justification.dart';
+export 'src/bundled_fonts.dart';